+++ packages/terra-framework-docs/src/terra-dev-site/test/table/ResizableColumnsTable.test.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import Table from 'terra-table';
+import tableData from './mockResizableColumnsTableData.json';
+
+const ResizableTable = () => {
+  const { cols, rows } = tableData;
+
+  return (
+    <Table
+      id="resizable-columns-table"
+      overflowColumns={cols}
+      rows={rows}
+      rowHeaderIndex={0}
+      columnHeaderHeight="50px"
+      ariaLabel="table"
+    />
+  );
+};
+
+export default ResizableTable;