+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 1.44.0 - (November 9, 2023)
 
 * Added
+  * Added examples and tests to `terra-table` for table sections.
   * Updated customDisplay examples for tabs.
   * Added `aria-label` and `ariaDescribedBy` examples for `terra-collapsible-menu-view`.
   * Added examples and test to `terra-table` for row selection mode.